Mega changes

+++ clients/Crm.Clients.Accounts/Models/AccountSettingType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Crm.Clients.Accounts.Models
 {
-    public enum AccountSettingType : short
+    public enum AccountSettingType : byte
     {
         PartnersEnabled = 1
     }
